discogs - fix (most of) release thumbnails, tests, cache
discogs apparently changed their api at the beginnig of march and we
were not able any more to get the albums thumbnails without an Oauth
authentication (as they explain in their doc). This patch includes
tests to reproduce the issue and a new method to get (most of) the thumbnails
from a different (discogs) source.