ui, deposit creation: reset card list on distributor selection change.
we would prefer to keep the cards that already have the right
distributor, but we don't have access to that information (yet ?).

use case: I select a distributor (or don't select any) and I begin to
select cards. Then if I change the distributor I would submit a form
with cards that don't belong to the selected distributor. This isn't a
logic pb (the necessary checks are done), it is a usability pb because I
would be distracted by error messages.