My Stock, New Entry: search for many ISBNs at once (¡only ones in stock!)
for #156

We can enter comma-separated ISBNs into the input field.
This returns the ISBNs we find in stock.

CAUTION: we don't currently search for unknown ISBNs in the remote
datasource. If we had a lot of ISBNs, this will take time, and making
the UI right is not trivial.
The middle ground is:
- with new books, one should scan them one by one into a List, or into
new Entry
- with books from the stock, one can use a memeroy barcode scanner and
search for many at once. Thus, we can still manipulate many at
once (add to a list, command, etc).

The current Inventory and List UIs are also not adapted for this.