Applying an inventory: set the distributor of *all* cards
scenario: we create a list which we leave empty, we choose its
supplier, and we "receive a command". We get an inventory view, we add
cards. When we applied the inventory, its cards were not updated with
the supplier linked to the basket. This is now fixed. Both the cards
of the inventory (or command received) as well as the cards that were
in the basket but not in the command have their supplier updated.