Create new basket: post params in request body
and not in url parameters. This made that a further click to increment
the card's quantity without a page refresh use this parameter,
contrarely to normal clicks. So the backend expected a json in the
request body, errored out, and the user was surprised to see the
quantity not incremeted in an export or after a page refresh.