Baskets JS: select input early if it is an ISBN
When we add a card to the list, the input is correctly emptied.
Here we select the ISBN in the field after it is fetched, before it is
validated.
Free text can not be selected at that moment or we wouldn't be able to
type more than 3 characters (the minimum set for the select2).