sell Bill: with list of books in the same document
for #99

When selling books, we can now:

- search for a client
  - or create a new one in a new window, and search for it in the sell
    window, without the need to refresh the page (select2 widget)
- create a bill in PDF for that client.

The bill contains: name and contact of us, the bookshop, name and
contact of the client, the total price before and after an optional
discount, the list of the sells (on another page), the payment
information (checks order, IBAN and BIC numbers).

Prior to that, one can enter the bookshop's personal and banking
information in the Preferences page.