basketToCommandController: augment page_size to 400
this actually limits the number of cards that are processed by the
validation (OK button) :S

with the speed up, we can increase this to a near-unreasonable amount.

If the command is bigger, the remaining cards will be kept in the "to
command" basket, and we can create a second command.

[ci skip]